backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
httpx==0.28.1
